Extract room screen action handling into separate component
